% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.PredictionCpt}
\alias{autoplot.PredictionCpt}
\title{Plots for Changepoint Predictions}
\usage{
\method{autoplot}{PredictionCpt}(object, task, type = "trace", theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{object}{(\code{PredictionCpt}).}

\item{task}{(\code{TaskCptUnsupervised} or \code{TaskCptSupervised})\cr
The task, required to extract the original data.}

\item{type}{(\code{character(1)})\cr
Type of the plot. See description.}

\item{theme}{(\code{ggplot2::theme()})\cr
The \code{ggplot2::theme_minimal()} is applied by default to all plots.}

\item{...}{(ignored).}
}
\value{
\code{ggplot2::ggplot()}.
}
\description{
Visualizations for \code{PredictionCpt}.
The argument \code{type} controls what kind of plot is drawn.
Possible choices are:
\itemize{
\item \code{"trace"} (default): Line plot of the target variable with vertical dashed lines representing the predicted changepoints.
\item \code{"segments"}: Step function representation of predicted changepoint segments.
}
}
