% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.TaskCptSupervised}
\alias{autoplot.TaskCptSupervised}
\title{Plots for Supervised Changepoint Tasks}
\usage{
\method{autoplot}{TaskCptSupervised}(object, type = "target", theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{object}{(\code{TaskCptSupervised}).}

\item{type}{(\code{character(1)})\cr
Type of the plot. See description.}

\item{theme}{(\code{ggplot2::theme()})\cr
The \code{ggplot2::theme_minimal()} is applied by default to all plots.}

\item{...}{(ignored).}
}
\value{
\code{ggplot2::ggplot()}.
}
\description{
Visualizations for \code{TaskCptSupervised}.
The argument \code{type} controls what kind of plot is drawn.
Possible choices are:
\itemize{
\item \code{"target"} (default): Line plot of the target variable along with vertical lines for the known changepoints.
}
}
