% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.TaskInterval}
\alias{autoplot.TaskInterval}
\title{Plots for Interval Regression Tasks}
\usage{
\method{autoplot}{TaskInterval}(object, theme = ggplot2::theme_minimal(), ...)
}
\arguments{
\item{object}{(\code{TaskInterval}).}

\item{theme}{(\code{ggplot2::theme()})\cr}

\item{...}{(ignored).}
}
\value{
\code{ggplot2::ggplot()}.
}
\description{
Visualizations for 'TaskInterval'
}
