% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3changepoint-package}
\alias{mlr3changepoint}
\alias{mlr3changepoint-package}
\title{mlr3changepoint: Weakly Supervised Change Point Detection for 'mlr3'}
\description{
Extends the 'mlr3' package with weakly supervised change point detection. Provides tasks, learners, and prediction objects for change point detection within the mlr3 framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sleepysnorlax05/mlr3changepoint}
  \item Report bugs at \url{https://github.com/sleepysnorlax05/mlr3changepoint/issues}
}

}
\author{
\strong{Maintainer}: Huy Nhat Minh Nguyen \email{sleepysnorlax0115@gmail.com}

Authors:
\itemize{
  \item Toby Hocking \email{toby.hocking@r-project.org}
  \item Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})
}

}
