% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCpt.R
\docType{class}
\name{PredictionCpt}
\alias{PredictionCpt}
\title{PredictionCpt}
\description{
This is the Prediction class for SUpervised and Unsupervised Change point detection, inheriting from \code{mlr3::Prediction}.
}
\section{Super class}{
\code{\link[mlr3:Prediction]{mlr3::Prediction}} -> \code{PredictionCpt}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{response}}{The predicted response}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionCpt-new}{\code{PredictionCpt$new()}}
\item \href{#method-PredictionCpt-clone}{\code{PredictionCpt$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="filter"><a href='../../mlr3/html/Prediction.html#method-Prediction-filter'><code>mlr3::Prediction$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="format"><a href='../../mlr3/html/Prediction.html#method-Prediction-format'><code>mlr3::Prediction$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="help"><a href='../../mlr3/html/Prediction.html#method-Prediction-help'><code>mlr3::Prediction$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="obs_loss"><a href='../../mlr3/html/Prediction.html#method-Prediction-obs_loss'><code>mlr3::Prediction$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="print"><a href='../../mlr3/html/Prediction.html#method-Prediction-print'><code>mlr3::Prediction$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="score"><a href='../../mlr3/html/Prediction.html#method-Prediction-score'><code>mlr3::Prediction$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionCpt-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionCpt-new}{}}}
\subsection{Method \code{new()}}{
Create a new PredictionCpt object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionCpt$new(
  task = NULL,
  row_ids = pdata$row_ids,
  response = pdata$response,
  pdata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\code{Task} The task for which predictions are made.}

\item{\code{row_ids}}{\code{integer} Row IDs corresponding to the predictions.}

\item{\code{response}}{\code{numeric} The predicted response (e.g., changepoint scores or binary labels).}

\item{\code{pdata}}{\code{list} or \code{PredictionData} Named list containing row_ids and response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionCpt-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionCpt-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionCpt$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
